+++ quasar-utils/src/main/groovy/net/corda/plugins/quasar/QuasarPlugin.groovy
@@ -39,7 +39,7 @@ class QuasarPlugin implements Plugin<Project> {
     private static final String CORDA_RUNTIME_ONLY_CONFIGURATION_NAME = 'cordaRuntimeOnly'
     @PackageScope static final String QUASAR_ARTIFACT_NAME = 'quasar-core'
     @PackageScope static final String defaultGroup = 'co.paralleluniverse'
-    @PackageScope static final String defaultVersion = '0.8.8_r3'
+    @PackageScope static final String defaultVersion = '0.8.9_r3'
 
     private final ObjectFactory objects